IMPROVE: autojoin関数のモジュール化
