+++ team2/client/src/features/cards/modules/AutoJoin.js
@@ -1,0 +1,23 @@
+export const AutoJoin = (scheculedTime, url) => {
+  const start_time = new Date(scheculedTime);
+  const now = new Date();
+  let is_started = false;
+
+  if(now <= start_time){
+    const autoJoinTimer = setInterval(() => {
+        const dt = new Date()
+        // console.log(dt);
+        if (dt.getMonth() === start_time.getMonth() &&
+            dt.getDate() === start_time.getDate() &&
+            dt.getHours() === start_time.getHours() &&
+            dt.getMinutes() === start_time.getMinutes()
+        ) {
+            window.open(url);//URL入れる
+            is_started = true;
+            clearInterval(autoJoinTimer);
+        }
+        // console.log('func; ' + is_started);
+    }, 60 * 1000);
+  }
+}
+
